extras/requirements.txt
# Compatible versions to avoid dependency conflicts
pdfplumber>=0.9.0
pandas>=1.5.0
matplotlib>=3.5.0
Pillow>=9.0.0
numpy>=1.22.4,<2.0.0  # Compatible with langchain, thinc, gensim

# OCR Libraries
pytesseract>=0.3.10
easyocr>=1.7.0
paddlepaddle>=2.4.0
paddleocr>=2.6.0

# Fast OCR alternatives (compatible versions)
transformers>=4.45.2,<5.0.0  # Compatible with marker-pdf
torch>=1.12.0
torchvision>=0.13.0

# Image preprocessing (avoid opencv conflicts)
opencv-python>=4.5.0,<4.11.0

# Optional: for GPU acceleration and faster inference
# accelerate>=0.20.0
# optimum>=1.8.0

# System requirements:
# For Tesseract OCR, also install:
# Ubuntu/Debian: sudo apt-get install tesseract-ocr tesseract-ocr-spa tesseract-ocr-eng
# macOS: brew install tesseract  
# Windows: https://github.com/UB-Mannheim/tesseract/wiki

# Installation note:
# If you have dependency conflicts, try:
# pip install --force-reinstall numpy==1.26.4
# Or create a new virtual environment